Vulkan: Fix bytesToAllocate calculation overflow bug
VertexArrayVk::updateStreamedAttribs() is calculating bytesToAllocate =
instanceCount * stride. The problem is that instanceCount and stride are
uint32 but bytesToAllocate is uint64. When the result of multiplication
exceeds UINT_MAX, it becomes zero which causes DynamicBuffer to crash.
This CL makes the calculation done in 64 bit math so that it will not
overflow. The buffer allocation code will fail when size exceeds its
limit and should handle in proper way.

Also Added a test that mimic webGL test with robust buffer access
enabled to demonstrate the crash in ANGLE vulkan backend.

Bug: angleproject:435228814
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/7012273
Commit-Queue: Charlie Lao <cclao@google.com>

+++ src/libANGLE/renderer/vulkan/VertexArrayVk.cpp
@@ -1361,8 +1361,10 @@ angle::Result VertexArrayVk::updateStreamedAttribs(const gl::Context *context,
             // Instanced attrib
             if (divisor > renderer->getMaxVertexAttribDivisor())
             {
-                // Divisor will be set to 1 & so update buffer to have 1 attrib per instance
-                size_t bytesToAllocate = instanceCount * stride;
+                // Divisor will be set to 1 & so update buffer to have 1 attrib per instance. Make
+                // sure we do multiplication with uint64_t in case the result is more than 32 bit.
+                // We will let buffer allocation code handle the failure if needed.
+                size_t bytesToAllocate = static_cast<size_t>(instanceCount) * stride;
 
                 // Allocate buffer for results
                 ANGLE_TRY(contextVk->allocateStreamedVertexBuffer(attribIndex, bytesToAllocate,
